fix deployment issue
