+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import ChatLayout from "@/components/chat/ChatLayout";
 import prisma from "@/lib/db";
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
-import { authOptions } from "./api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 
 export default async function Home() {
   const session = await getServerSession(authOptions)
